Small fixes and reformatting
git-svn-id: https://scm.ti-edu.ch/reposvn/opensmt2/opensmt-dev/trunk@634 d17752b6-1f7d-11e2-a792-abcf70c528dd

+++ src/options/SMTConfig.h
@@ -415,7 +415,7 @@ public:
   int sat_clause_decay() const
     { return optionTable.contains(o_clause_decay) ?
         optionTable[o_clause_decay]->getValue().decval : 1 / 0.999; }
-  int sat_random_var_freq() const
+  double sat_random_var_freq() const
     { return optionTable.contains(o_random_var_freq) ?
         optionTable[o_random_var_freq]->getValue().decval : 0.02; }
   int sat_random_seed() const
